app/requirements.txt
fastapi==0.81.0
uvicorn[standard]==0.18.3
websocket-client==1.5.0